render 함수를 메인페이지로 이동하고, init으로 네이밍 수정
